Added CoreMessageService
