+++ src/NuGetGallery.Core/packages.config
@@ -1,5 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
+  <package id="AnglicanGeek.MarkdownMailer" version="1.2" targetFramework="net452" />
   <package id="elmah.corelibrary.strongname" version="1.2.2" targetFramework="net452" />
   <package id="EntityFramework" version="6.1.3" targetFramework="net452" />
   <package id="Microsoft.Azure.KeyVault.Core" version="1.0.0" targetFramework="net452" />
