+++ tests/NuGetGallery.Core.Facts/NuGetGallery.Core.Facts.csproj
@@ -139,6 +139,7 @@
     <Compile Include="Entities\PackageFacts.cs" />
     <Compile Include="Entities\UserFacts.cs" />
     <Compile Include="Entities\UserSecurityPolicyFacts.cs" />
+    <Compile Include="Extensions\UserExtensionsFacts.cs" />
     <Compile Include="PackageStatusFacts.cs" />
     <Compile Include="Packaging\ManifestValidatorFacts.cs" />
     <Compile Include="Packaging\PackageIdValidatorTest.cs" />
